apps/prover/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
orjson==3.10.7
python-chess==1.999
pydantic==2.9.2
python-dotenv==1.0.1
garaga
